requirements.txt
# 智能问数功能依赖项

# 向量数据库
chromadb>=0.4.0

# 图数据库和图算法
networkx>=3.0
graphiti-core>=0.24.0

# LangChain 相关
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.8
langchain-experimental>=0.0.15

# OpenAI SDK（用于直接调用API）
openai>=1.0.0

# LangGraph 工作流框架
langgraph>=0.0.40

# 数据库连接
pymysql>=1.1.0
psycopg2-binary>=2.9.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 中文分词
jieba>=0.42.0

# 环境变量管理
python-dotenv>=1.0.0

# 多智能体框架（圆桌会议第一/二/三层智能体可选用）
agentscope>=1.0.0
